Make collections into composable
